updated code

+++ src/main/java/com/jahcodework/universal_pet_care/repository/VetRepo.java
@@ -3,5 +3,5 @@ package com.jahcodework.universal_pet_care.repository;
 import com.jahcodework.universal_pet_care.model.Vet;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VetRepo extends JpaRepository<Vet, Integer> {
+public interface VetRepo extends JpaRepository<Vet, Long> {
 }
